+++ src/com/company/test.java
@@ -1,4 +1,15 @@
 package com.company;
 
+import javax.swing.*;
+import java.awt.*;
+
 public class test {
+    private JPanel panel1;
+    private JPanel panel2;
+    private JPanel panel3;
+
+    private void createUIComponents() {
+        // TODO: place custom component creation code here
+    }
+
 }
